modern: use mongoose for mongodb connection; add dotenv; move remaining routes to routes/
