+++ modules/timestamp.js
@@ -1,0 +1,3 @@
+export function timestamp() {
+	return `[${new Date().toLocaleTimeString('en-GB')}]`
+}
