+++ package.json
@@ -15,11 +15,13 @@
 	"license": "ISC",
 	"dependencies": {
 		"body-parser": "^1.19.0",
+		"dotenv": "^16.0.3",
 		"ejs": "^3.1.6",
 		"express": "^4.17.1",
 		"express-session": "^1.17.1",
 		"geolib": "^3.3.1",
 		"mongodb": "^4.0.0-beta.3",
+		"mongoose": "^6.8.0",
 		"prettier": "^2.8.1",
 		"superagent": "^6.1.0"
 	},
